Fix: DTOs unificados

+++ Producto/backend/src/main/java/backend/com/shared/domain/model/Pais.java
@@ -9,7 +9,7 @@ import lombok.*;
 @Builder
 public class Pais {
 
-    private Integer idPais;
+    private Long idPais;
     private String nombrePais;
 
 }
